available-videos: move streamVideo functions to available-videos
